좌상 로직 변경, views와 service 분리
