+++ qldb/services/qldb_setting.py
@@ -1,0 +1,83 @@
+
+# from boto3 import client
+import boto3
+from logging import basicConfig, getLogger, INFO
+from qldb.ledger import *
+from pyqldb.driver.qldb_driver import QldbDriver
+import uuid
+
+from amazon.ion.simple_types import IonPyBool, IonPyBytes, IonPyDecimal, IonPyDict, IonPyFloat, IonPyInt, IonPyList, \
+    IonPyNull, IonPySymbol, IonPyText, IonPyTimestamp
+from amazon.ion.simpleion import dumps, loads
+from django.conf import settings
+from time import sleep
+
+
+
+
+logger = getLogger(__name__)
+basicConfig(level=INFO)
+IonValue = (IonPyBool, IonPyBytes, IonPyDecimal, IonPyDict, IonPyFloat, IonPyInt, IonPyList, IonPyNull, IonPySymbol,
+            IonPyText, IonPyTimestamp)
+
+ledger_name=QLDB.LEDGER_NAME
+
+access_key=getattr(settings,'AWS_ACCESS_KEY_ID')
+secret_access_key=getattr(settings,'AWS_SECRET_ACCESS_KEY')
+region_name=getattr(settings,'AWS_REGION')
+
+session=boto3.session.Session(aws_access_key_id=access_key, 
+                              aws_secret_access_key=secret_access_key, 
+                              region_name=region_name)
+
+qldb_client = session.client(
+    'qldb'
+    )
+
+qldb_driver = QldbDriver(
+    ledger_name=ledger_name, 
+    boto3_session=session
+    )
+
+LEDGER_CREATION_POLL_PERIOD_SEC = 10
+ACTIVE_STATE = "ACTIVE"
+
+
+
+
+def create_qldb_ledger(name):
+
+    logger.info("Let's create the ledger named: {}...".format(name))
+    result = qldb_client.create_ledger(Name=name, PermissionsMode='ALLOW_ALL')
+    logger.info('Success. Ledger state: {}.'.format(result.get('State')))
+    return result
+
+
+def wait_for_active(name):
+ 
+    logger.info('Waiting for ledger to become active...')
+    while True:
+        result = qldb_client.describe_ledger(Name=name)
+        if result.get('State') == ACTIVE_STATE:
+            logger.info('Success. Ledger is active and ready to use.')
+            return result
+        logger.info('The ledger is still creating. Please wait...')
+        sleep(LEDGER_CREATION_POLL_PERIOD_SEC)
+        
+def create_qldb_table(driver, table_name):
+    
+    logger.info("Creating the '{}' table...".format(table_name))
+    statement = 'CREATE TABLE {}'.format(table_name)
+    cursor = driver.execute_lambda(lambda executor: executor.execute_statement(statement))
+    logger.info('{} table created successfully.'.format(table_name))
+    # print(len(list(cursor)))
+    return len(list(cursor))
+
+
+def create_table_index(driver, table_name, index_attribute):
+    
+    logger.info("Creating index on '{}'...".format(index_attribute))
+    statement = 'CREATE INDEX on {} ({})'.format(table_name, index_attribute)
+    cursor = driver.execute_lambda(lambda executor: executor.execute_statement(statement))
+    return len(list(cursor))    
+
